fix: Cast interval days calculation to int in OphTrOperationbooking_Operation_Sequence model
The getDateListForMonths method was dividing interval by 24*60*60 which resulted in a float, but mktime() expects an integer for the day parameter. This was causing 'Implicit conversion from float to int loses precision' errors in PHP 8.1+.

+++ protected/modules/OphTrOperationbooking/models/OphTrOperationbooking_Operation_Sequence.php
@@ -371,7 +371,7 @@ class OphTrOperationbooking_Operation_Sequence extends BaseActiveRecordVersioned
                 $interval = $interval_obj ? $interval_obj->getInteger($endDate) : null;
             }
 
-            $days = $interval ? $interval / 24 / 60 / 60 : 0;
+            $days = $interval ? (int)($interval / 24 / 60 / 60) : 0;
 
             $nextStartDate = $startDate;
